Defines whether the cursor should keep its current position.
Defines whether the cursor should keep its position when
text gets inserted at the current position of the cursor.

For example, we don't want QTextCursor to extend the selection
when inserting characters at the end of an extra selection
representing a user-type but we do want the selection to grow
when rewriting a region of code (e.g. when using our QuickFix engine).

Task-number: QTBUG-11075
Reviewed-by: mae