Clipping with rounded rectangles and QVG_SCISSOR_CLIP
WebKit draws rounded rectangles by setting a rounded rect path as a
clip and then filling the clip.  This had problems when QVG_SCISSOR_CLIP
was enabled because the scissor would only clip to the bounding box
of complex paths.  Thus, rounded rectangle clips were turned into
normal rectangle clips.

This change will make the scissor clipping code subdivide complex paths
into rectangles and use a tighter clip if the number of rectangles is
less than or equal to VG_MAX_SCISSOR_RECTS.  If it is greater, then it
will fall back to the bounding box of the path.

Task-number: QT-3372
Reviewed-by: Jason Barron