Fix Qt.sis content for Symbian^3 builds
Some build-time omissions from Qt_template.pkg are done, so qt.sis
built on Symbian^3 will not work properly on most S60 3.x & 5.0 devices.
This should not be an issue, as Symbian^3 builds are expected to be
configured to use OpenVG anyway.

If necessary, it is still possible to build Qt.sis compatible with
older platforms also on Symbian^3 if you fake the S60 version when
calling qmake, e.g.: 'qmake S60_VERSION=5.0'

Reviewed-by: Shane Kearns