Fix the bindTexture() filters on X11 to not rely on mipmap generation.
The default GL_TEXTURE_MIN_FILTER is GL_NEAREST_MIPMAP_LINEAR, which
means you have to have mipmaps generated in order for it to produce
correct results. We don't use mipmaps when this particular path
is taken, so use texture filters that is in line with what we do in the
non-pixmap_to_texture path.

Task-number: QTBUG-11158
Reviewed-by: Kim