QWidget::childAt for masked child widgets doesn't work properly
Problem was that we didn't take the children's effective mask
into account when checking whether a point was inside or not.

This commit is also an optimization since we no longer check the point
against the widget's rect twice. Furthermore, unnecessary
QRegion operations are completely avoided in the common case.

Auto tests included.

Task-number: QTBUG-7150
Reviewed-by: paul