Retrieve proper glyph outlines from Symbian^4
Up until Symbian^4, the Symbian font Api does not provide access to the
vector data of glyphs. In order to make QPainterPath::addText() and
transformed text rendering work, Qt falls back to a workaound: Tracing
the bitmap of a rasterized glyph. That is neither performant nor is the
result of good quality.

Symbian^4 will add Api that returns glyph outlines encoded as SVG paths.
The code in this commit essentially parses the SVG path and returns a
QPainterPath. The parsing part took some ideas from QSvg, and is
implemented in a very reduced way.

Task-number: QT-809
Reviewed-by: Aleksandar Sasha Babic