New keys for media-related keys, common mobile keys
Various new keys:

- introduce a proper key for pausing media - the existing Key_Pause was
mapped to the Pause/Break key on a QWERTY keyboard.  (In order not to break
BC on applications that have been using the directfb keyboard map, the
key mapping for DIKS_PLAYPAUSE *was not* updated).
- introduce explicit keys to handle the case where play/pause is a single
(toggling state) key rather than two discrete keys; same thing for a
toggling version of call/hangup.
- add new keys for voice dial and last number redial

NOTE: Due to renumbering Key_Camera and Key_CameraFocus in order to make
space in the more mobile-specific key range, this change introduces a
binary compatibility break between this change and anything built before
(ie. with Qt 4.7 beta 1 or earlier)
(But not Qt 4.6, since the camera and camera focus keys were not in 4.6)

Reviewed-by: axis
Reviewed-by: Shane Kearns