Disabled item view items use incorrect background color
This was a problem when using alternate row colors.
We use a separate palette for base/alternate when
using disabled colors. However, we should only use
this if the entire view is disabled. To keep compatibility
with style sheets we have to preserve the disabled
state per item, but we now use the widget pointer to
decide which palette role to use.

Task-number: QTBUG-11263
Reviewed-by: ogoffart