Fixed crash which occurs when switching between video clips
When Phonon::MediaObject::setCurrentSource() is called when the
MediaObject is connected to a Phonon::VideoWidget, the
MMF::AbstractVideoOutput pointer is propagated inside the backend
from the first MMF::AbstractVideoPlayer to the second.

If the VideoWidget is subsquently re-sized, the code path enters
the ScaleFactors branch of the
MMF::SurfaceVideoPlayer::handleParametersChanged function.  At this
point, m_displayWindow is still set to the inital null value, and the
assertion therefore fails.

This change ensures that m_displayWindow is updated before attempting
to apply the scale factor change.

Task-number: QTBUG-11377
Reviewed-by: Thierry Bastian