Fixed `nmake clean' breaking declarative imports on Windows.
Generally, doing a `clean' ought to clean up space by deleting
intermediate objects, but leave the build in a working state
(as opposed to `distclean' which deletes intermediate and target
objects).

For this to work for declarative imports, the qmldir file should
be treated as a target, not an intermediate object.