Fix recursive 'make sis' for symbian builds
It makes no sense for 'sis' to depend on 'first' or really any target
in symbian-abld and symbian-sbsv2 builds, as all builds are recursive
anyway.

Also do not generate real sis targets for projects that do not
actually deploy anything, making the check equivalent to one
in qmake for generating pkg files.

Task-number: QTBUG-5312
Reviewed-by: Janne Koskinen