Thread safety for QFontEngineS60
On symbian, creating fonts requires a connection to the window server.
Window server sessions are not sharable across threads.
To avoid QFont crashing when used outside the GUI thread, we construct a
private session to the window server when the shared CONE session is not
available.
(CCoeEnv::Static() uses TLS, so it returns null outside of GUI thread)

The private session and screen device are stored in QThreadStorage, so
they are automatically deleted when the QThread exits.

Task-number: QTBUG-8874
Reviewed-by: mread