Add QTextOption API to QStaticText
To allow more control over the layout process for a QStaticText,
introduce a QTextOption API. This will allow you to, e.g., set an
alignment for the text layout. Patch also contains a couple of bug fixes
to make right alignment work correctly when you set a text width on
the QStaticText.

Task-number: QTBUG-9031
Reviewed-by: Samuel