fix detection of header files
Disable the include generation iff a dot which is not followed by [hH]
is found after the last path separator. This implies that dots in
directory names are now ignored, and that files without an extension are
always considered headers (e.g., STL headers and Qt forwarding headers).

Task-number: QTBUG-11369
Merge-request: 686
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>