Fix regression, remove artificial emboldening of monospaced fonts
In change 889e13652e00ba64060e7b62ed3377ea04d44faa we employed usage
of artificial emboldening of fonts with FreeType when a bold variant
of the font does not exist. This introduced a regression for monospaced
fonts, where the contract says that all glyphs must be the same width.
Since the artificial emboldening algorithm does not support this
contract, it has been turned off for monospaced fonts, reverting to
the old behavior for this particular class of fonts.

Task-number: QTBUG-11353
Reviewed-by: Paul