Fix some kind of race condition while using remote commands.
Please do not merge to master, we had to fix it different there.

Because of async content loading, we have to cache the url to load. A
combination of SetSource and SyncContents would have lead to sync to the
old url till loading was finished. So return the cached during loading.

Task-number: QTBUG-11342
Reviewed-by: ck