Make sure only started gestures can cause cancellations
The design is that calling setGestureCancelPolicy on the gesture
that is in starting state, and is accepted can cause other gestures
to be cancelled.
So change the logic to follow this, which means we won't try to execute
the cancel strategy if the gesture was never accepted and avoid the
crash detailed in the task.

Reviewed-by: Denis
Task-number: QTBUG-9771