Fix event forwarding in QDeclarativeFlickable.
The flickable element filters all events of its children and store the
press event to replay it if there is a release or if the scrolling
didn't happen. The issue was that the event and the item stored to
"replay" the press event might not be the item that is interessted by
the event. Let say you have a translucent overlay on top of an other
item. Previously all events will be send to the overlay and not to the
item underneath. This happen beause QGraphicsView propagate events from
top to bottom (stacking order) so the overlay will be the first child
filtered by the flickable. So we need to repropagate the event through
the normal process to the event delivery mechanism of QGraphicsView
will work properly. Also we need to unset the mouse grabber since after
the first press it might be set to a wrong item. We also need to replay
the release by ourself on the new mouse grabber but only if we need to
send again the press.

Reviewed-by:Yann Bodson