QLocalSocket/Win: handle ERROR_MORE_DATA after read operation
If we're connected to a name pipe which is in message mode, we have to
handle the following case: ReadFile() or GetOverlappedResult() return
FALSE and GetLastError() returns ERROR_MORE_DATA.

This just means, that the message didn't fit into the pipe's internal
buffer. We must not handle this as error.

Task-number: QTBUG-11490
Reviewed-by: ossi