Rework the internal GL resource system yet again.
Instead of having to sub-class QGLContextResource, we now have two
template classes:

1. 'QGLContextResource' to handle context specific resources that are
*not* shared in a group, even though the context itself is in a sharing
group

2. 'QGLContextGroupResource' to handle resources shared in a context
group.

Classes used as the template class type must have a constructor with
the following signature:
T(const QGLContext *);

The resources that the templates wrap are freed up when the context
or context group it is tied to is destroyed. The resource is also
freed when the QGLContext<Group>Resource object is destroyed, which
might lead to context switches, depending on where and in how many
non-sharing contexts the resource is used.

The templates wrap some boiler plate code that were common for all the
use cases.