Start the implementation of in-place recoding for images
Currently, with the graphics system raster, converting from images
to QPixmap often needs to allocate a new image to convert the right
format.

For example, for an image in ARGB32 of 10 mbytes, we need to allocate
a second image of 10 mbytes in ARGB32_PM to convert the source image
in the right format for pixmap.

This can create a hight peak of memory, and is a bit slower
than it should.

This patch introduce in-place conversion of images when they are
loaded with QPixmap::loadFromData().
The images are loaded in their default format by QImageReader,
and are then converted in-place, trying to reduce memory
allocations.

Reviewed-by: Samuel Rødal