Prevent warnings about EPOCROOT coming out when not on symbian
When the pro file is processed with qmake and it EPOCROOT is not set
then it will give out warnings.  This is really only an issue when
configure on X11 is ran because this runs qmake on all pro files
anyway.

So rather than confusing users about what this warning message means on
a non Symbian platform, we just ensure that the pro file specifically
only does anything useful on Symbian platforms.

Reviewed-by: Gareth Stockwell