Fix a crash on exit. Destruction order in font database. (Symbian^4)
The implementation for the QT_2746 task had a wrong destruction order
of font database elements. The 'QSymbianTypeFaceExtras' in Symbian^4
hold a CFont which is retrieved from the Symbian ScreenDevice. They
have to be released by the same ScreenDevice. The error was that
the release was attempted after the connection to the ScreenDevice
was closed.

This fix causes an earlier destruction of all 'QSymbianTypeFaceExtras'
and the removal of their their CFonts in qt_cleanup. While the
connection to the ScreenDevice is still alive.

Task-number: QT_2746