Fix painting of overlines in labels.
If the size of the widget is an odd number, the layout rect
will be positioned at a 0.5 offset. The clipping set further
below will snap this to 1 and thus clip away the anything drawn
on the first row of pixels. Fix this by aligning the rectangle
we get by layout rect.

Reviewed-by: Eskil
Task: http://bugreports.qt.nokia.com/browse/QTBUG-9684