Enablers for external use of QGLTextureGlyphCache
In order to enable use of QGLTextureGlyphCache outside of Qt, it can't
depend on having an initialized paint engine. We use the simple image
fallback for copying the cache texture when it's updated and we need
to export the functions for resolving extensions, so that the gl context
can be initialized properly without calling begin on a paint engine.

Reviewed-by: Kim