Revert "Fixing the race condition in event dispatcher implementation on"
This reverts commit 5f21450a61ba2459e6dc5b08b236b15a067bf81f.

It's the wrong fix for the problem. The socket operations in the
select thread are not supposed to be thread safe. It is thread safe
between the select thread and the caller thread, but not if there are
more than one caller thread. Ensuring complete thread safety would
probably require fixes in other areas as well.

The correct fix is for the client to ensure that calls to the socket
registering functions are either serialized or contained within one
thread.

Task:     QT-3358
RevBy:    Aleksandar Sasha Babic