Revert "Fixing race condition in qeventdispatcher_symbian.cpp code path"
This reverts commit 3a9ca8ba50a63e564c15023f65cbac36d365d309.

It's the wrong fix for the problem. The socket operations in the
select thread are not supposed to be thread safe. It is thread safe
between the select thread and the caller thread, but not if there are
more than one caller thread. Ensuring complete thread safety would
probably require fixes in other areas as well.

The correct fix is for the client to ensure that calls to the socket
registering functions are either serialized or contained within one
thread.

Task:     QT-3496
RevBy:    Aleksandar Sasha Babic