No more enter & leave events after a popup menu is closed on Windows
We use the TrackMouseEvent() to keep track of enter and leave messages.
Upon receiving a WM_MOUSELEAVE, the effect of that function call is over,
we need to call it again if we need more messages. This was not done when
we close a popup because of mouse click outside.

Task-number: QTBUG-11582
Reviewed-by: Bradley T. Hughes
(cherry picked from commit d806b996beab76cbc43ffa11bc38721fa19c9c81)