Layout direction change by key should change alignment of QLineEdit
When QLineControl was split out of QLineEdit, a regression was
introduced. When the layout direction was altered by a key press,
the layout direction of the QLineControl would be set, but not the
QLineEdit. This would in turn mean that QLineEdit would use the wrong
layout direction for its visual alignment. Patch is a small hack to
read back the layout direction from the control after it has been
set.

Task-number: QTBUG-11204
Reviewed-by: Trond
(cherry picked from commit d198186cdb18bbb9428ca51e8771cc4fd9563452)