Add support for comments and meta-data in the lupdate QtScript frontend
This brings the support more in line with the C++ frontend.
No support for magic TRANSLATOR comment yet.
Add a proper test for the QtScript parser.
Support for qtTrId() and friends coming in a separate commit.

Previously the lexer just skipped comments. Now the contents
of each comment is retained and passed to a "comment processor"
interface. The parser implements the interface and performs
logic very similar to the C++ parser to extract data from
the comment.

Task-number: QTBUG-11774
Reviewed-by: Oswald Buddenhagen