Streamline lupdate's QtScript frontend's error messaging
Like the C++ parser, use a yyMsg() function.
Make the lexer a member of the parser, so yyMsg() can access
the filename and current line number.
This refactoring is done in preparation of two changes that
introduce quite a few more potential error message calls
(comment processing and qtTrId support).

Reviewed-by: Oswald Buddenhagen