SSL backend: correct ordering of root CA certificates in the store
Some CAs have several root certificates with the same SubjectInfo, but
different expiration date. This means that OpenSSL needs to choose the
valid root certificate over the expired ones when verifying a
certificate chain. Apparently, the certs added to the store first are
the ones that get used first, so we add the expired ones at the end.

Reviewed-by: Thiago Macieira
Task-number: QTBUG-11664