Fix RVCT compile error in QGraphicsSceneIndex autotest
The test code used the 'using' keyword to try and change access control
of a base class method from protected to public.
With the RVCT 2.2 compiler, 'using' imports the function(s) from the base
class, but they retain their existing access control.

Used an inline public function to call the base class as a workaround

Reviewed-by: mread