ColorDialog is incorrectly positioned
Due to somewhat special nature of colorDialog, it makes no sense to
make it fullscreen, as it is not totally stretchable in height. Making
it fullscreen only makes it appear higher on the screen and there is
an ugly gap between the dialog and softkeys.

Therefore, as a fix, colorDialog is shown in its "natural" window
state.

Task-number: QTBUG-11668
Reviewed-by: Miikka Heikkinen