Crash when dragging with mingw
The problem was that there was a misalignment coming from the d&d code.
That called in the end our SSE optimized functions.
This code turns out to be calling QApplication::processEvents which
in the the calls the Qt event dispatcher for windows' processEvents
function. This function will now also always align the stack to 16
bytes and be SSE ready.

Task-number: QTBUG-11880
Reviewed-by: Zeno Albisser