Fix crash when handleCommand() called before softkeys are updated
Softkeys are updated via a compressable event that is posted via the
event loop. Since these events are not delivered immediately, there is
a chance that a call to handleCommand() could happen before the
softkeys have been updated which can lead to a crash if the previous
QAction's have been deleted already since the data structure used by
QSoftKeyManager is outdated. The likeliness of this is increased by the
fact that S60 commands are normally sent from the WSERV active object
which has a higher priority than the active object used by Qt's event
loop which means commands will preempt the event loop.

The fix is to introduce a flag that keeps track of pending update
requests and if a command is received while there are outstanding
requests, force the softkeys to be updated before handling the command.

Task-number: QT-3571
Reviewed-by: axis