Warn when drawPixmapFragments is called with an invalid source rect
When drawPixmapFragments() is called with fragments that has invalid
source rects in it, then usually it causes the pixmap drawn on screen
to appear corrupted.  However it has been reported that a crash can
occur (not reproducable locally) so by adding a warning in debug mode
only means that this can hopefully be caught at development time.

Reviewed-by: Trond