Minor modifications and optimizations to the Symbian HW glyph cache.
After the latest round of API reviews, the Symbian glyph cache has now
been moved from GDI to FBS and this required some changes in Qt's
implementation.

Also incorporate an optimization where we first iterate over the glyph
vector to eliminate glyphs that are already cached in Qt's glyph cache.
This way we only open the glyph iterator on glyphs that we need.

Reviewed-by: Alessandro Portale