Added QDeclarativeSpringAnimation
The QDeclarativeSpringAnimation is a replacement for
QDeclarativeSpringFollow. The idea is to remove the Follows
quickly.

Follows used to have an inSync property. In order to provide
an alternative mechanism, the commit also fixes the
running property for animations which are controlled by a
behavior. Previously running would always return false and
never change. Now running does change and indicates that the
animation is running indeed.