Don't call GetDC() nedlessly.
It seems that the CS_OWNDC class style flag only returns the same
DC when GetDC() is called in the same thread that created the window.
This in turn generates a bad DC resource leak if you have a rendering
thread with one or more GL context switches, so just call GetDC() when
you actually need to.

Reviewed-by: Kim