Make Text, TextInput, and TextEdit all have the same size for the same text. This may mean that the cursor is to the right of the width, so components should cater for that with a margin if they are boxed / clipped.
TextInput used to try to account for right bearing (and left bearing incorrectly since
it added it on the right). This is removed. Potentially this means that for some fonts
the text repaints incorrectly on the left or right, but if that is the case Text and TextEdit
already had such a problem (undetected), and all will need fixing.

Task-number: QTBUG-11983