Set QThreads to be process critical automatically on Symbian OS
On Symbian OS, a thread is non critical by default - this means if it
crashes the process does not automatically terminate, rather it is
allowed to handle the crash by itself. Only the main thread is critical
by default.
Since this is not the behaviour on other platforms, application developers
are unlikely to have considered this case - and would need to write
symbian specific code to catch the crash.

For such advanced users, they can reset the critical flag on the thread
once they get control in QThread::run().
By default, a crash in any thread created via the QThread API will now
crash the process.

Reviewed-by: Iain