Fix subfocus when reparenting panels.
When reparenting an item into another, QGraphicsItem will fix the
subfocus chain by first removing subfocus from the previous ancestors,
and then optionally restoring the subfocus chain to the new ancestors
of the reparented item. This allows item trees to be created and
assigned a specific focus item before even added to a scene.

The QGraphicsItemPrivate::setSubFocus function did not take into
account that the root item might be in a different panel than the
reparented item, which is the case when reparenting one panel into
another. If so, setSubFocus must not mess up the new ancestor panel's
sub focus chain. Otherwise (former behavior), the new ancestor
panel's focusItem will point to the reparented panel's focus item. This
in turn breaks an invariant; one panel's focusItem can never point to
an item in another panel.

The bundled test case shows this point: Create two panels that each have
one child that sets focus. The first panel is added to the scene just
to prepare, and verify that it's activated as expected, and its focus
child immediately gains input focus. Then we create a second panel,
which also has a focus child (but it doesn't have focus as its panel is
not active). When we reparent the new panel onto the old one, we expect
the new panel to be activated, and the child focus item to gain input
focus. However, we also expect the parent panel's focus item to remain
intact, so that when we hide the child panel, and the parent panel is
reactivated, its original focus item regains focus.

Former behavior was that the parent's focus item still pointed to the
child panel focus item.

This change only has effect when reparenting to or from a panel.

Merge-request: 735
Reviewed-by: Yoann Lopes <yoann.lopes@nokia.com>