Add support for -runtimegraphicssystem configure option
It has been requested that we provide a configuration option to specify
the default "inner" graphics system that should be used when the
runtime graphics system is in use. The argument specified to the option
is written to the qconfig.h as QT_DEFAULT_RUNTIME_SYSTEM and this is
used to instantiate the default graphics system that will be used by
the runtime graphics system.

Reviewed-by: Gunnar Sletta