Optimize text layout.
When laying out text, a significant amount of time was being spent
querying for the font engine. This patch:
* changes the layout to only query for the engine when a new script item
is encountered.
* adds an internal cache of the previous result to
QTextEngine::fontEngine(). This catches the important case of multiline
text with few font engine changes.

With these changes layout costs are now approximately 60% of what they
were previously, as measured by the text layout benchmarks.

Reviewed-by: Eskil Abrahamsen Blomfeldt