Fix USB serial port detection of the Nokia N95 on linux
The interface numbers in /dev/serial/by-id are hex rather than decimal.
Also added code to read the manufacturer name and product name from string descriptors
in order to get a better match. Unfortunately, root privilege is needed or the API
returns an error.
In this case, we still use the weak matching on interface number only.

Task-Number: QTBUG-11794
Reviewed-By: Thomas Zander