Linux runonphone - tell the user which driver to load
The linux usbserial driver doesn't attach to devices automatically, to avoid conflicts
with the real driver (if there is one). So the user must run modprobe to load the
driver before runonphone can be used.
Since runonphone has found the correct device(s) using libusb, this patch will tell the
user the modprobe command that is required to load the generic usbserial driver.

Reviewed-By: Gareth Stockwell