Fixes crash in QGraphicsScene::addItem().
Crashed because tabFocusFirst could end up being a dangling pointer when
removing an item from the scene before deleting it.
When setting tabFocusFirst in fixFocusChainBeforeReparenting, we now
check that the item is in the scene. If it is not, tabFocusFirst is set
to 0.

Autotest included.

Task-number: QTBUG-12056
Reviewed-by: Alexis Ménard <alexis.menard@nokia.com>
(cherry picked from commit 7d09f690d5b4c56699092444665d1879deb86a6e)