Fix QTextDocument::markContentsDirty()
The function markContentsDirty(from,length) is documented
to inform the document that it needs to lay out again
the part from from to from + length.

Trouble was that the function was implemented using
beginEditBlock()/endEditBlock(), which has the negative
side effect that it does increase the document's revision
number and emits all sorts of contentChanged() signals.

This did not matter from the one case where Qt uses the
method itself in QSyntaxHighlighter, because here a private
flag inContentsChange avoided the issue.

Without the change, we cannot implement semantic syntax
highlighting in creator without expensive extra selections
or triggering a full rehighlight via QSyntaxHighlighter.

Done-with: Roberto Raggi
Reviewed-by: Roberto Raggi