Make QFontEngineGlyphCache explicitly shared
QFontEngine will delete its glyph cache (and if it's a GL engine, its
texture) when the font engine dies. The font engine can die when the
QFontCache decides to clean it up, e.g. based on the trigger of a timer.
To make it possible to use the glyph cache elsewhere, without
interference by the font cache, it has been made explicitly shared.

Reviewed-by: Kim