syncqt: fix wrong paths in include/ActiveQt/headers.pri
Assigning a semicolon separated list of directories to a Qt module does not
work correctly. The values of %modules must be split into single values
before operating on them. This doesn't happen in all places.

ActiveQt is the only module where this feature is used.
Also, it is not needed anymore for this particular module.
Thus, I'll just remove the semicolon separated list from ActiveQt and
replace it with the module directory.

Task-number: QTBUG-4586
Reviewed-by: ossi