Support text decoration in QPainter::drawGlyphs()
In order to support text decorations (overline, underline and strikeout)
in QPainter::drawGlyphs(), we need to call drawTextItemDecoration()
from this function. To support this, it has been generalized to no
longer require a QTextItemInt. We also need to propagate the decoration
attributes of the font from QTextLayout to QGlyphs, since this is not
part of the QFontEngine.

Task-number: QTBUG-12122
Reviewed-by: Kim